Another major update
- incorporate some of Ravi's suggestions
- the intro section is being pared down a lot, with the same material
being instead covered in a new "Problem" section, which will also argue
for the value-add of the approach so that we maybe don't need a separate
"Evaluation" section. The Problem section isn't finished yet, but this
commit gives a clear outline of what it will look like and where it is going.