change function names in code
`lkup` is more consistent with the agda, and i think a bit more
intuitive, than `look`

I think `uncons` is misleading - although it's technically accurate,
given the implementation, we don't want to leak these implementation
details through the interface. The client doesn't need to know that the
underlying impl is a list, so we are only saying that we are pulling out
an arbitrary mapping, not the first mapping in particular. Instead,
calling it `destruct` makes it clear that we will eliminate a single
mapping, but not specify anything about which particular mapping we are
pulling out. Also, the term `destruct` is consistent with the rest of
the paper.