Rewrite case study, making it totally concrete
We may revisit some of the stuff mentioned in the previous case study.
It made some additional arguments for the practical value of the
properties, going beyond the cases where the properties are absolutely
necessary, and pointing out how they could make proofs easier or less
awkward in scenarios where they're not strictly necessary. I think this
is useful content, esp. regarding the costs of refinement proofs and
concordant benefit of semantically total data structures.