New case study section elaborating practical benefits of Totality
This stuff is borrowed from a previous version of the case study - it
talks about how even if we don't upset the positivity checker,
refinement proofs are still a liability for a lot of day to day
programming, due to proof relevance and related issues. I think it's
good to have this here, to make a broader statement about the value of
semantic totality and why it's nice to not have to refine your type
unless you really really need to.